+++ pkg/httpclient/client_test.go
@@ -39,6 +39,25 @@ func TestQueryTrace(t *testing.T) {
 		assert.True(t, proto.Equal(trace, response))
 	})
 
+	t.Run("includes the recentDataTarget header", func(t *testing.T) {
+		mockTransport := MockRoundTripper(func(req *http.Request) *http.Response {
+			assert.Equal(t, liveStoreHeaderValue, req.Header.Get(recentDataTargetHeader))
+			response, _ := proto.Marshal(trace)
+			return &http.Response{
+				StatusCode: 200,
+				Body:       io.NopCloser(bytes.NewReader(response)),
+			}
+		})
+
+		client := New("www.tempo.com", "1000")
+		client.QueryLiveStores = true
+		client.WithTransport(mockTransport)
+		response, err := client.QueryTrace("100")
+
+		assert.NoError(t, err)
+		assert.True(t, proto.Equal(trace, response))
+	})
+
 	t.Run("returns a trace not found error on 404", func(t *testing.T) {
 		mockTransport := MockRoundTripper(func(_ *http.Request) *http.Response {
 			return &http.Response{
@@ -56,7 +75,7 @@ func TestQueryTrace(t *testing.T) {
 	})
 }
 
-func TestQueryTraceWithRance(t *testing.T) {
+func TestQueryTraceWithRange(t *testing.T) {
 	trace := &tempopb.Trace{}
 	t.Run("returns an error if start time is greater than end time", func(t *testing.T) {
 		client := New("www.tempo.com", "1000")